change rate, phase back to doubles. improves read/write position drift

+++ src/SoftCutVoice.cpp
@@ -123,7 +123,7 @@ void SoftCutVoice::setFilterFcMod(float x) {
 }
 
 void SoftCutVoice::updateFilterFc() {
-    float fc = std::min(fcBase, fcBase * std::fabs(sch.getRate()));
+    float fc = std::min(fcBase, fcBase * std::fabs(static_cast<float>(sch.getRate())));
     // std::cout << fc << std::endl;
     svf.setFc(fc*fcMod + (1.f-fcMod )*svf.getFc());
 }
